Moves Number class into separate file
Preparation for actually splitting the class into two.
