+++ lib/src/core/reader.dart
@@ -4,6 +4,7 @@ import 'dart:convert' show JSON;
 import 'dart:math' show min;
 
 import 'expressions.dart';
+import 'numbers.dart';
 import 'project_interface.dart';
 
 Expression schemeRead(List<Expression> tokens, ProjectInterface impl) {
